backend/requirements.txt
Flask>=2.2.5,<3.0
Flask-Login>=0.6.2
Flask-SQLAlchemy>=3.1.1
openai
python-dotenv>=1.0.0
Werkzeug>=2.2.2,<3.0
Flask-Cors>=4.0.0
Flask-Bcrypt